+++ Source/WinCopies.Collections.Shared/Enumeration/Enumerators/Enumerator.cs
@@ -46,12 +46,7 @@ namespace WinCopies.Collections
                 throw GetExceptionForDispose(false);
         }
 
-        protected T GetOrThrowIfDisposed<T>(in T value) =>
-#if WinCopies3 && CS9
-            ((WinCopies.DotNetFix.IDisposable)this).GetOrThrowIfDisposed(value);
-#else
-            WinCopies.ThrowHelper.GetOrThrowIfDisposed(this, value);
-#endif
+        protected T GetOrThrowIfDisposed<T>(in T value) => WinCopies.ThrowHelper.GetOrThrowIfDisposed(this, value);
 
         public bool MoveNext()
         {